+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.1.0
+
+- add: settings for turning on saturation and/or lightness sliders
+- add: sliders for saturation and lightness
+- fix: landscape mode
+
 ## 1.0.1+1
 
 - fix: maxColors can be null